+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityXpBottle.java
@@ -1,7 +1,5 @@
 package org.allaymc.api.entity.interfaces;
 
-import org.allaymc.api.entity.Entity;
-
-public interface EntityXpBottle extends Entity {
+public interface EntityXpBottle extends EntityProjectile {
 
 }