+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemExperienceBottleStack.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemProjectileComponent;
 
-public interface ItemExperienceBottleStack extends ItemStack {
+public interface ItemExperienceBottleStack extends ItemStack, ItemProjectileComponent {
 }