+++ server/src/main/java/org/allaymc/server/entity/impl/EntityXpBottleImpl.java
@@ -1,14 +1,27 @@
 package org.allaymc.server.entity.impl;
 
-import java.util.List;
+import lombok.experimental.Delegate;
 import org.allaymc.api.component.interfaces.Component;
+import org.allaymc.api.entity.component.EntityAgeComponent;
+import org.allaymc.api.entity.component.EntityPhysicsComponent;
+import org.allaymc.api.entity.component.EntityProjectileComponent;
 import org.allaymc.api.entity.initinfo.EntityInitInfo;
 import org.allaymc.api.entity.interfaces.EntityXpBottle;
 import org.allaymc.server.component.interfaces.ComponentProvider;
 
+import java.util.List;
+
 public class EntityXpBottleImpl extends EntityImpl implements EntityXpBottle {
+
+    @Delegate
+    protected EntityProjectileComponent projectileComponent;
+    @Delegate
+    protected EntityPhysicsComponent physicsComponent;
+    @Delegate
+    protected EntityAgeComponent ageComponent;
+
     public EntityXpBottleImpl(EntityInitInfo initInfo,
-            List<ComponentProvider<? extends Component>> componentProviders) {
-        super(initInfo, componentProviders);;
+                              List<ComponentProvider<? extends Component>> componentProviders) {
+        super(initInfo, componentProviders);
     }
 }