+++ server/src/main/java/org/allaymc/server/item/impl/ItemExperienceBottleStackImpl.java
@@ -1,14 +1,20 @@
 package org.allaymc.server.item.impl;
 
-import java.util.List;
+import lombok.experimental.Delegate;
 import org.allaymc.api.component.interfaces.Component;
+import org.allaymc.api.item.component.ItemProjectileComponent;
 import org.allaymc.api.item.initinfo.ItemStackInitInfo;
 import org.allaymc.api.item.interfaces.ItemExperienceBottleStack;
 import org.allaymc.server.component.interfaces.ComponentProvider;
 
+import java.util.List;
+
 public class ItemExperienceBottleStackImpl extends ItemStackImpl implements ItemExperienceBottleStack {
+    @Delegate
+    protected ItemProjectileComponent projectileComponent;
+
     public ItemExperienceBottleStackImpl(ItemStackInitInfo initInfo,
-            List<ComponentProvider<? extends Component>> componentProviders) {
+                                         List<ComponentProvider<? extends Component>> componentProviders) {
         super(initInfo, componentProviders);
     }
 }