+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - Introduced component `EntityBreatheComponent` which handle the logic of entity breathe.
 - Added armor component to the newly added copper armor items.
 - Introduced a new optional arg for command `/enchant` which indicates whether the entered level should be checked to be in acceptable range.
+- Implemented xp bottle.
 
 ### Fixed